+++ CHANGELOG.md
@@ -1,6 +1,13 @@
 ## 0.7.4
 - Fix return value boxing of equals and not equals operators
 - Fix assigning enums in conditional expressions
+- Fix error thrown by null assertion operator (thanks @Noobware1)
+- Add bindings for `Error`, `UnimplementedError`, `UnsupportedError`, 
+  and `StackTrace` (thanks @Noobware1)
+- Add bindings for `InternetAddress` and `InternetAddressType` (thanks 
+  @Noobware1)
+- Support for calling `toString()` on dynamic types (thanks @Noobware1)
+- Support for `jsonEncode()` and `jsonDecode()` (thanks @Noobware1)
 
 ## 0.7.3
 - Fix error when setting a top-level variable